msal: Move user reqs from misc to msal::user
